Refactor UI project to MVVM structure
