+++ UI.LoraSort/Converters/IndexToPriorityConverter.cs
@@ -2,7 +2,7 @@
 using System.Globalization;
 using System.Windows.Data;
 
-namespace UI.LoraSort
+namespace UI.LoraSort.Converters
 {
     public class IndexToPriorityConverter : IMultiValueConverter
     {
