+++ UI.LoraSort/Views/Dialog/EditMappingWindow.xaml
@@ -1,7 +1,7 @@
-﻿<Window x:Class="UI.LoraSort.NewMappingWindow"
+﻿<Window x:Class="UI.LoraSort.Views.Dialog.EditMappingWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="New Mapping" Height="200" Width="400" WindowStartupLocation="CenterOwner">
+        Title="Edit Mapping" Height="200" Width="400" WindowStartupLocation="CenterOwner">
     <Grid Margin="10">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
