+++ UI.LoraSort/Views/MainWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="UI.LoraSort.MainWindow"
+﻿<Window x:Class="UI.LoraSort.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:UI.LoraSort.ViewModels"
-        xmlns:conv="clr-namespace:UI.LoraSort"
-        xmlns:controls="clr-namespace:UI.LoraSort"
+        xmlns:conv="clr-namespace:UI.LoraSort.Converters"
+        xmlns:controls="clr-namespace:UI.LoraSort.Views.UserControls"
         mc:Ignorable="d"
         Title="Lora Auto Sort - By AIKnowledge2Go" Height="800" Width="1000"
         MinHeight="900" MinWidth="1200"
