+++ UI.LoraSort/Views/MainWindow.xaml.cs
@@ -2,7 +2,9 @@
 using System.Windows;
 using UI.LoraSort.ViewModels;
 
-namespace UI.LoraSort
+using UI.LoraSort.Views.Dialog;
+using UI.LoraSort.Services;
+namespace UI.LoraSort.Views
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
