+++ UI.LoraSort/Views/UserControls/CustomMappingsTopRightControl.xaml
@@ -1,7 +1,7 @@
-﻿<UserControl x:Class="UI.LoraSort.CustomMappingsTopRightControl"
+﻿<UserControl x:Class="UI.LoraSort.Views.UserControls.CustomMappingsTopRightControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:conv="clr-namespace:UI.LoraSort"
+             xmlns:conv="clr-namespace:UI.LoraSort.Converters"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
