+++ UI.LoraSort/Views/UserControls/ProcessLogControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="UI.LoraSort.ProcessLogControl"
+﻿<UserControl x:Class="UI.LoraSort.Views.UserControls.ProcessLogControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
