+++ UI.LoraSort/Views/UserControls/SettingsControl.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="UI.LoraSort.UserControls.SettingsControl"
+<UserControl x:Class="UI.LoraSort.Views.UserControls.SettingsControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
